Rename is-primitive to isPrimitive.
