+++ lib/Model.js
@@ -17,7 +17,6 @@ var arraySlice = require("./support/array-slice");
 var collectLru = require("./lru/collect");
 var pathSyntax = require("falcor-path-syntax");
 
-var isPrimitive = require("./support/is-primitive");
 var isJsonEnvelope = require("./support/is-json-envelope");
 var isJsonGraphEnvelope = require("./support/is-json-graph-envelope");
 
@@ -27,6 +26,7 @@ var getSize = require("./support/getSize");
 var isObject = require("./support/isObject");
 var isFunction = require("./support/isFunction");
 var isPathValue = require("./support/isPathValue");
+var isPrimitive = require("./support/isPrimitive");
 var jsong = require("falcor-json-graph");
 var ID = 0;
 var validateInput = require("./support/validate-input");
