+++ lib/support/isFunction.js
