+++ lib/support/isObject.js
