+++ lib/support/isPathValue.js
@@ -1,5 +1,5 @@
 var isArray = Array.isArray;
-var isObject = require("./../support/is-object");
+var isObject = require("./../support/isObject");
 
 module.exports = function isPathValue(pathValue) {
     return isObject(pathValue) && (
