+++ lib/support/isPrimitive.js
